requirements.txt
Flask==1.1.4
msal==1.16.0
python-dotenv==0.19.2
requests==2.26.0
Werkzeug==1.0.1
flask-cors==3.0.10